aws-greengrass/services/sam-emergency/src/requirements.txt
awsiotsdk
numpy
